api/requirements.txt
alembic==1.8.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
filelock==3.8.0
Flask==2.1.0
Flask-JWT==0.3.2
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.0.1
gunicorn==20.0.4
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.5
PyJWT==1.4.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.3.19
urllib3==1.26.12
virtualenv==20.16.6
Werkzeug==2.0.0
WTForms==3.0.1
WTForms-JSON==0.3.5
zipp==3.10.0
